requirements.txt
appdirs==1.4.4
audioread==2.1.9
azure-core==1.8.2
azure-storage-blob==12.5.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
decorator==4.4.2
dtw-python==1.1.6
idna==2.10
isodate==0.6.0
joblib==0.17.0
librosa==0.8.0
llvmlite==0.34.0
msrest==0.6.19
numba==0.51.2
numpy==1.19.3
oauthlib==3.1.0
packaging==20.4
pandas==1.1.3
pika==1.1.0
pooch==1.2.0
PyAudio==0.2.11
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
resampy==0.2.2
scikit-learn==0.23.2
scipy==1.5.3
six==1.15.0
SoundFile==0.10.3.post1
threadpoolctl==2.1.0
urllib3==1.25.11
